+++ tests/command_line.sh
@@ -1,6 +1,8 @@
 #!/bin/bash
 chunkflow generate-tasks -c 0 0 0 -s 0 0 0 -g 1 1 1
 
+chunkflow create-chunk plugin -f median_filter
+
 chunkflow create-chunk -o seg create-chunk -o gt evaluate-segmentation -s seg -g gt
 
 chunkflow log-summary -l tests/data/log