+++ CHANGELOG.md
@@ -2,11 +2,15 @@ ChangeLog history
 =================
 # chunkflow 1.0.1 (2020-xx-xx)
 ## Breaking Changes
-
+- renamed custom-operator to plugin
 ## Deprecations 
 
 ## Features
 - new operator: normalize-intensity
+- a plugin system with a median filter example
+- new operator: normalize-intensity
+- support cutout from hdf5 file
+
 ## Bug Fixes 
 - cutout whole volume in default