+++ CHANGELOG.md
@@ -6,8 +6,10 @@ ChangeLog history
 ## Deprecations 
 
 ## Features
+- new operator: normalize-intensity
 - a plugin system with a median filter example
 - new operator: normalize-intensity
+- support cutout from hdf5 file
 
 ## Bug Fixes 
 - cutout whole volume in default